split out common subroutines
